+++ python-lib/dku_aws/eksctl_command.py
@@ -1,6 +1,6 @@
 import sys, os, subprocess, logging, json, requests, shutil
 from .eksctl_loader import get_eksctl_or_fetch
-from dku_utils.access import _has_not_blank_property
+from dku_utils.access import _has_not_blank_property, _convert_to_string
 
 class EksctlCommand(object):
     def __init__(self, args, connection_info):
@@ -11,11 +11,13 @@ class EksctlCommand(object):
             self.env['AWS_ACCESS_KEY_ID'] = connection_info['accessKey']
         if _has_not_blank_property(connection_info, 'secretKey'):
             self.env['AWS_SECRET_ACCESS_KEY'] = connection_info['secretKey']
+        if _has_not_blank_property(connection_info, 'sessionToken'):
+            self.env['AWS_SESSION_TOKEN'] = connection_info['sessionToken']
         if _has_not_blank_property(connection_info, 'region'):
             self.env['AWS_DEFAULT_REGION'] = connection_info['region']
         
     def run(self):
-        cmd = [self.eksctl_bin] + self.args
+        cmd = _convert_to_string([self.eksctl_bin] + self.args)
         logging.info('Running %s' % (' '.join(cmd)))
         p = subprocess.Popen(cmd,
                              shell=False,
@@ -26,12 +28,12 @@ class EksctlCommand(object):
         (o, e) = p.communicate()
         rv = p.wait()
         return (cmd, rv, o, e)
-    
+
     def run_and_get_output(self):
         return self.run()[2]
     
     def run_and_log(self):
-        cmd = [self.eksctl_bin] + self.args
+        cmd = _convert_to_string([self.eksctl_bin] + self.args)
         logging.info('Running %s' % (' '.join(cmd)))
         p = subprocess.Popen(cmd,
                              shell=False,
@@ -41,11 +43,11 @@ class EksctlCommand(object):
                              universal_newlines=True)
         with p.stdout as s:
             for line in iter(s.readline, ''):
-                logging.info(line)
+                logging.info(line.rstrip())
         return p.wait()
     
     def run_and_get(self):
-        cmd = [self.eksctl_bin] + self.args
+        cmd = _convert_to_string([self.eksctl_bin] + self.args)
         logging.info('Running %s' % (' '.join(cmd)))
         p = subprocess.Popen(cmd,
                              shell=False,