+++ python-runnables/add-node-pool/runnable.py
@@ -5,9 +5,8 @@ from dku_kube.autoscaler import add_autoscaler_if_needed
 from dku_kube.gpu_driver import add_gpu_driver_if_needed
 from dku_aws.eksctl_command import EksctlCommand
 from dku_aws.aws_command import AwsCommand
-from dku_utils.cluster import get_cluster_from_dss_cluster
-from dku_utils.access import _has_not_blank_property
-from dku_utils.config_parser import get_security_groups_arg
+from dku_utils.cluster import get_cluster_from_dss_cluster, get_connection_info
+from dku_utils.config_parser import get_security_groups_arg, get_region_arg
 
 class MyRunnable(Runnable):
     def __init__(self, project_key, config, plugin_config):
@@ -32,7 +31,7 @@ class MyRunnable(Runnable):
         # the cluster is accessible via the kubeconfig
         kube_config_path = dss_cluster_settings.get_raw()['containerSettings']['executionConfigsGenericOverrides']['kubeConfigPath']
 
-        connection_info = dss_cluster_config.get('config', {}).get('connectionInfo', {})
+        connection_info = get_connection_info(dss_cluster_config)
         
         node_group_id = self.config.get('nodeGroupId', None)
         
@@ -42,10 +41,7 @@ class MyRunnable(Runnable):
         if node_group_id is not None and len(node_group_id) > 0:
             args = args + ['--name', node_group_id]
         
-        if _has_not_blank_property(connection_info, 'region'):
-            args = args + ['--region', connection_info['region']]
-        elif 'AWS_DEFAULT_REGION' is os.environ:
-            args = args + ['--region', os.environ['AWS_DEFAULT_REGION']]
+        args = args + get_region_arg(connection_info)
             
         if dss_cluster_config['config'].get('useEcr', False):
             args = args + ['--full-ecr-access']
@@ -85,10 +81,7 @@ class MyRunnable(Runnable):
         #args = args + ['-v', '4']
         args = args + ['--cluster', cluster_id]
 
-        if _has_not_blank_property(connection_info, 'region'):
-            args = args + ['--region', connection_info['region']]
-        elif 'AWS_DEFAULT_REGION' is os.environ:
-            args = args + ['--region', os.environ['AWS_DEFAULT_REGION']]
+        args = args + get_region_arg(connection_info)
 
         args = args + ['-o', 'json']